Move redis client to own file
