+++ redis/index.go
@@ -1,0 +1,30 @@
+package redis
+
+import (
+	"context"
+	"time"
+
+	"github.com/go-redis/redis/v8"
+)
+
+type Client struct {
+	client *redis.Client
+}
+
+func NewRedis() (*Client, error) {
+	client := redis.NewClient(&redis.Options{
+		Addr:        "localhost:6379",
+		Password: "",
+		DB:          0,
+		DialTimeout: 100 * time.Millisecond,
+		ReadTimeout: 100 * time.Millisecond,
+	})
+
+	if _, err := client.Ping(context.Background()).Result(); err != nil {
+		return nil, err
+	}
+
+	return &Client{
+		client: client,
+	}, nil
+}
